+++ pyArango/connection.py
@@ -54,6 +54,7 @@ class AikidoSession(object):
                 retry = 0
                 while status_code == 1200 and retry < self.max_conflict_retries :
                     ret = self.fct(*args, **kwargs)
+                    status_code = ret.status_code
                     retry += 1
             except:
                 print ("===\nUnable to establish connection, perhaps arango is not running.\n===")
@@ -67,7 +68,7 @@ class AikidoSession(object):
             ret.json = JsonHook(ret)
             return ret
 
-    def __init__(self, username, password, max_conflict_retries=5, single_session=True, verify=True, max_retries=5, log_requests=False):
+    def __init__(self, username, password, verify=True, max_conflict_retries=5, max_retries=5, single_session=True, log_requests=False):
         if username:
             self.auth = (username, password)
         else:
@@ -122,7 +123,38 @@ class AikidoSession(object):
 class Connection(object):
     """This is the entry point in pyArango and directly handles databases.
     @param arangoURL: can be either a string url or a list of string urls to different coordinators
-    @param use_grequests: allows for running concurent requets."""
+    @param use_grequests: allows for running concurent requets.
+
+    Parameters
+    ----------
+    arangoURL: list or str
+        list of urls or url for connecting to the db
+
+    username: str
+        for credentials
+    password: str
+        for credentials
+    verify: bool
+        check the validity of the CA certificate
+    verbose: bool
+        flag for addictional prints during run
+    statsdClient: instance
+        statsd instance    
+    reportFileName: str
+        where to save statsd report
+    loadBalancing: str
+        type of load balancing between collections
+    use_grequests: bool
+        parallelise requests using gevents. Use with care as gevents monkey patches python, this could have unintended concequences on other packages
+    use_jwt_authentication: bool
+        use JWT authentication
+    use_lock_for_reseting_jwt: bool
+        use lock for reseting gevents authentication
+    max_retries: int
+        max number of retries for a request
+    max_conflict_retries: int
+        max number of requests for a conflict error (1200 arangodb error). Does not work with gevents (grequests),
+    """
 
     LOAD_BLANCING_METHODS = {'round-robin', 'random'}
 
@@ -139,6 +171,7 @@ class Connection(object):
         use_jwt_authentication=False,
         use_lock_for_reseting_jwt=True,
         max_retries=5,
+        max_conflict_retries=5
     ):
 
         if loadBalancing not in Connection.LOAD_BLANCING_METHODS:
@@ -151,6 +184,7 @@ class Connection(object):
         self.use_jwt_authentication = use_jwt_authentication
         self.use_lock_for_reseting_jwt = use_lock_for_reseting_jwt
         self.max_retries = max_retries
+        self.max_conflict_retries = max_conflict_retries
         self.action = ConnectionAction(self)
 
         self.databases = {}
@@ -230,7 +264,16 @@ class Connection(object):
                 verify
             )
         else:
-            self.session = AikidoSession(username, password, verify, self.max_retries)
+            # self.session = AikidoSession(username, password, verify, self.max_retries)
+            self.session = AikidoSession(
+                username=username,
+                password=password,
+                verify=verify,
+                single_session=True,
+                max_conflict_retries=self.max_conflict_retries,
+                max_retries=self.max_retries,
+                log_requests=False
+            )
 
     def reload(self):
         """Reloads the database list.