+++ CHANGELOG.rst
@@ -4,6 +4,9 @@
 * restoreIndex and restoreIndexes in collection will restore previously deleted indexes
 * added max_conflict_retries to handle arango's 1200
 * added single session so AikidoSessio.Holders can share a single request session
+* added task deletion to tests reset
+* added drop() to tasks to remove all tasks in one command 
+* better documentation of connection class
 
 1.3.4
 =====